requirements.txt
torch>=2.1.0
transformers>=4.38.0
pennylane>=0.37.0
datasets>=2.18.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
tqdm>=4.66.0
gradio>=5.0.0
pyyaml>=6.0.1
sentencepiece>=0.2.0
protobuf>=4.25.0
accelerate>=0.29.0
